committing all code to main
This repo doesn't have a good base of updates because I didn't clone the repository I forked it. Essentially, this is a copy of another repository because I'm a doofus. I'll link that repository
https://github.com/danpatoneil/Work-Day-Scheduler-fake
so you can look at the commits from there, but it wouldn't let me publish from that repo